requirements.txt
aiofiles
appdirs
APScheduler
arabic-reshaper
asgiref
astroid
attrs
autobahn
Automat
autopep8
black
blinker
certifi
cffi
channels
chardet
click
constantly
cryptography
daphne
dataclasses
defusedxml
dj-database-url
Django
django-allauth
django-apscheduler
django-common-helpers
django-cors-headers
django-crispy-forms
django-cron
django-crontab
django-easy-pdf
django-email-verification
django-guardian
django-jsonfield
django-rest-passwordreset
django-six
django-taggit
django-truncate
django-xhtml2pdf
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
environ
future
geographiclib
geopy
googlemaps
gunicorn
h11
h2
hpack
html5lib
Hypercorn
hyperframe
hyperlink
idna
incremental 
isort 
itsdangerous
Jinja2 
lazy-object-proxy 
MarkupSafe 
mccabe 
mypy-extensions 
numpy 
oauthlib 
pandas
pathspec 
pep8 
Pillow
priority 
pyasn1 
pyasn1-modules 
pycodestyle 
pycparser
PyHamcrest 
PyJWT 
pylint 
pyOpenSSL 
PyPDF2 
python-bidi 
python-dateutil 
python3-openid 
pytz 
Quart 
regex 
reportlab 
requests 
requests-oauthlib 
ruamel.yaml 
ruamel.yaml.clib
service-identity 
six 
sqlparse 
toml 
Twisted 
txaio 
typed-ast 
typing-extensions 
tzlocal 
urllib3 
webencodings 
Werkzeug
whitenoise 
wrapt 
wsproto
xhtml2pdf
zope.interface
